update on orders

+++ api/v1/views/company.py
@@ -43,32 +43,33 @@ def get_company():
 @app_views.route("/company", methods=["PUT"], strict_slashes=False)
 @jwt_required()
 def update_company(id: str = None):
-    try:
-        comp_id = get_jwt_identity()
-        if not comp_id:
-            return jsonify(not_found), 404
-        company = storage.get(Company, comp_id)
-        if company:
-            up_data = request.form.to_dict()
-            if 'image' in request.files and request.files['image']:
-                up_data['image'] = save_file(request.files['image'])
+    #try:
+    comp_id = get_jwt_identity()
+    if not comp_id:
+        return jsonify(not_found), 404
+    company = storage.get(Company, comp_id)
+    if company:
+        up_data = request.form.to_dict()
+        if 'image' in request.files and request.files['image']:
+            up_data['image'] = save_file(request.files['image'])
+            if company.image:
                 image_file = os.path.basename(company.image)
                 if image_file:
                     delete_file(image_file)
-            for key, value in up_data.items():
-                if key == "__class__" or key == "id" or  key ==\
+        for key, value in up_data.items():
+            if key == "__class__" or key == "id" or  key ==\
                     "created_date" or key == "updated_date":
-                    pass
-                else:
-                    setattr(company, key, value)
-            company.save()
-            return jsonify(company.to_dict())
-        else:
-            print(" data from", up_data)
-            return jsonify(not_found), 404
-    except Exception as e:
-        print(e)
-        return jsonify(error_data), 505
+                pass
+            else:
+                setattr(company, key, value)
+        company.save()
+        return jsonify(company.to_dict())
+    else:
+        print(" data from", up_data)
+        return jsonify(not_found), 404
+    # except Exception as e:
+    #     print(e)
+    #     return jsonify(error_data), 505
 
 
 @app_views.route("/company/<id>", methods=["DELETE"], strict_slashes=False)
